wordpress-microfrontend/requirements.txt
Flask==2.3.3
Werkzeug>=2.3.7
Jinja2>=3.1.2
itsdangerous>=2.1.2
click>=8.1.3
blinker>=1.6.2

opentelemetry-api==1.36.0
opentelemetry-sdk==1.36.0
opentelemetry-semantic-conventions==0.57b0
opentelemetry-exporter-otlp==1.36.0
opentelemetry-instrumentation==0.57b0
opentelemetry-instrumentation-flask==0.57b0
opentelemetry-instrumentation-wsgi==0.57b0
